apps/local-agent/pyproject.toml
[tool.poetry]
name = "phi-agent"
version = "0.1.0"
description = "Local agent runtime for Phi Agents"
authors = ["Phi Agents Team"]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
httpx = "^0.25.2"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
click = "^8.1.7"

[tool.poetry.scripts]
phi-agent = "phi_agent.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


